Introducing New Zephyr: a complete rewrite :^)
Zephyr was build very incrementally, adding new features started bacome
more and more difficult as the overall architecture was not always made
by taking into account the whole picture. Now that I have a good idea of
what I need I'm confident that a complete rewrite is the right thing to
do.

Also, there are few things I would like to do that were not possible
with the previous architecture, the most notable being switching to
incremental compilation and going much deeper in the "query oriented"
design. I expect that this will make the compiler usefull not only to
get fast "code -> compile -> debug" cycles but also as a fundation for
more advance tooling such as IDE, formatter, linter...